feat: Auth 유틸 통합
