+++ lib/auth.ts
@@ -1,5 +1,30 @@
+import { NamuiError } from '@/error'
+
 export type Provider = 'kakao'
-export const getNewToken = async () => '11'
+export const getNewToken = async () => new NamuiError.NotImplimentError()
+
+export interface SignInOptions {
+  callbackUrl?: string
+  redirect?: boolean
+}
+
+export async function signIn(
+  provider: Provider,
+  options?: SignInOptions,
+): Promise<void> {
+  const { callbackUrl = window.location.href, redirect = true } = options ?? {}
+  console.log(callbackUrl, '<<')
+
+  const url = Oauth.getAuthorizationURL(provider)
+  if (!url) {
+    throw new NamuiError.BadRequestError()
+  }
+  if (redirect) {
+    window.location.href = url.toString()
+    if (url.toString().includes('#')) window.location.reload()
+    return
+  }
+}
 
 export class Oauth {
   static getAuthorizationURL(provider: Provider | string) {
